Inter-layer connections wrapping & input template gen
Connections between the different layers of the arm are made. Separate distances for the layers can be defined, as well as separate materials. 

There is also a function created that will output a template image to apply force to the wrapped arm. 

Creating supports and translating image to force has to be done still.